requirements.txt
alabaster==0.7.12
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
commonmark==0.9.1
docutils==0.16
et-xmlfile==1.0.1
idna==2.10
imagesize==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
openpyxl==3.0.5
packaging==20.4
Pillow==8.0.1
Pygments==2.7.2
pyhunter==1.7
pyparsing==2.4.7
pytz==2020.4
recommonmark==0.6.0
requests==2.25.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.2
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.3.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.26.2